+++ TextRpg/Player/Mage.cs
@@ -11,7 +11,7 @@ namespace TextRpg.Player
         // 체력 100, 공격력 15, 방어력 5 
         // 장비 장착 설정 시 아래 참고
         // string id, string name, int level, int exp, float maxexp, int Str, int Agi, int Int, int hp, int mp, float itematk, float itemdef, int gold, int weapon, int armor, List<int> items
-        public Mage(string id, string name, int level, int exp, float maxexp, int Str, int Agi, int Int, int hp, int mp, int gold, List<int> items) : base(id, name, "마법사", level, exp, maxexp, Str, Agi, Int, hp, mp, 0, 0, gold, 0, 0, items, false)
+        public Mage(string id, string name, int level, int exp, int maxexp, int Str, int Agi, int Int, int hp, int mp, int gold, List<int> items) : base(id, name, "마법사", level, exp, maxexp, Str, Agi, Int, hp, mp, 0, 0, gold, 0, 0, items, false)
         {
             if (Health <= 0)
             {
